refactor(analytics): move analytics files to `app/core/analytics/`
