+++ app/core/analytics/models.py
